+++ crates/tests/misc/calling_convention/tests/win.rs
@@ -5,7 +5,7 @@ use windows::Win32::*;
 fn calling_convention() {
     unsafe {
         // This function requires cdecl on x86.
-        assert_eq!(LdapMapErrorToWin32(LDAP_BUSY as u32), ERROR_BUSY);
+        assert_eq!(LdapMapErrorToWin32(LDAP_BUSY as u32), ERROR_BUSY as u32);
 
         // This function requires stdcall on x86.
         GetTickCount();
