+++ crates/tests/misc/win32/tests/winsock.rs
@@ -80,7 +80,7 @@ fn in_addr() {
     fn in_addr_to_string(in_addr: &IN_ADDR) -> String {
         unsafe {
             let mut chars = [0u8; 24];
-            inet_ntop(AF_INET as i32, in_addr as *const IN_ADDR as _, &mut chars);
+            inet_ntop(AF_INET, in_addr as *const IN_ADDR as _, &mut chars);
             std::ffi::CStr::from_ptr(chars.as_ptr() as *const _)
                 .to_str()
                 .unwrap()
